feat: implement POM service helpers
